wado-compiler: extract resource_rewrite submodule from cm_binding
Move the CM resource method rewriting pass and the per-record
stream-read binding generator into
`synthesis/cm_binding/resource_rewrite.rs`. Both run before adapter
synthesis.
